+++ User.md
@@ -1,0 +1,5 @@
+- User login/create/fetchOne/fetchAll
+
+* Create User Schema with commonSchema-
+* Define User router within User module
+* Defining controller for each functionality and link it to router
