+++ demo/utils/commonSchema.js
@@ -1,0 +1,14 @@
+const mongoose = require('mongoose');
+
+const commonSchema = new mongoose.Schema({
+  createdAt: {
+    type: Date,
+    default: Date.now,
+  },
+  updatedAt: {
+    type: Date,
+    default: Date.now,
+  },
+});
+
+module.exports = { commonSchema };
