+++ server/src/modules/user.controller.js
@@ -1,0 +1,27 @@
+import User from "./user.model";
+
+const getAllUser = (req, res) => {
+  return "getallUser";
+};
+
+const getUser = (req, res) => {
+  return "getUser";
+};
+
+const updateUser = (req, res) => {
+  return "updateUser";
+};
+
+const deleteUser = (req, res) => {
+  return "deleteUser";
+};
+
+const loginUser = (req, res) => {
+  return "loginUser";
+};
+
+const createUser = (req, res) => {
+  return "createUser";
+};
+
+export { getAllUser, getUser, updateUser, deleteUser, loginUser, createUser };
