--- server/utils/validateEmail.js
@@ -1,6 +1,0 @@
-const validateEmail = (email) => {
-  const emailRegex = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/;
-  return emailRegex.test(email);
-};
-
-module.exports = validateEmail;
